modificada base de datos

+++ expense_tracker/db/migrate/20150513203127_create_transactions.rb
@@ -1,9 +1,9 @@
 class CreateTransactions < ActiveRecord::Migration
-	def change
+  def change
 		create_table :transactions do |t|
 			t.date :date
 			t.integer :t_type, limit: 1
-			t.float :amount
+			t.decimal :amount
 			t.string :description, limit: 50
 			
 			t.references :category, index: true, foreign_key:true
